Move Dependency class out of the way, fixes #11
It seems the <dependency> tag in the requireSameVersions rules tries to
use this class whenever it is present in the same package. Moving it out
of the way fixes that problem.
