+++ src/it/classname-conflict/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals=enforcer:enforce
